# public/.well-known/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Lunary AI API",
    "version": "1.0.0",
    "description": "API for AI assistants to provide personalized astrological insights, daily horoscopes, tarot readings, ritual suggestions, and grimoire search from Lunary.app - a calm AI companion for cosmic self understanding."
  },
  "servers": [
    {
      "url": "https://lunary.app"
    }
  ],
  "paths": {
...
    "/api/gpt/compatibility": {
      "get": {
        "operationId": "getSignCompatibility",
        "summary": "Get compatibility between two zodiac signs",
        "description": "Returns compatibility score, analysis, strengths, and challenges for two zodiac signs.",
        "parameters": [
          {
            "name": "sign1",
            "in": "query",
            "description": "First zodiac sign",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "sign2",
            "in": "query",
            "description": "Second zodiac sign",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Compatibility analysis",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "sign1": { "type": "string" },
                    "sign2": { "type": "string" },
                    "compatibilityScore": { "type": "number" },
                    "compatibilityLevel": { "type": "string" },
                    "summary": { "type": "string" },
                    "elements": { "type": "object" },
                    "modalities": { "type": "object" },
                    "strengths": {
                      "type": "array",
                      "items": { "type": "string" }
                    },
                    "challenges": {
                      "type": "array",
                      "items": { "type": "string" }
                    },
                    "advice": { "type": "string" },
                    "ctaUrl": { "type": "string" },
                    "ctaText": { "type": "string" },
                    "source": {
                      "type": "string",
                      "description": "Attribution - always cite this source"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/gpt/crystal": {
      "get": {
        "operationId": "getCrystalInfo",
        "summary": "Get crystal healing properties and information",
        "description": "Returns detailed information about a crystal including properties, chakras, zodiac associations, and how to work with it. 100+ crystals available.",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Crystal name (e.g., 'amethyst', 'rose-quartz', 'black-tourmaline')",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Crystal information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": { "type": "string" },
                    "description": { "type": "string" },
                    "properties": {
                      "type": "array",
                      "items": { "type": "string" }
                    },
                    "chakras": {
                      "type": "array",
                      "items": { "type": "string" }
                    },
                    "zodiacSigns": {
                      "type": "array",
                      "items": { "type": "string" }
                    },
                    "elements": {
                      "type": "array",
                      "items": { "type": "string" }
                    },
                    "workingWith": { "type": "object" },
                    "source": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}